winui3: Phase 1i tools + reviewer fixes
Lands slice 1i (ToolResolver / tools.ts port) plus the BLOCKER and
IMPORTANT findings from the cross-runtime GPT-5.5 + Opus-4.8 rubber-duck
review of slices 1a/1b/1g/1h.

Slice 1i  Ccw.Core/Tools
* ToolResolver.cs ports src/tools.ts: candidate-list path resolution
  for npm, atk, the M365 Agents Toolkit binary, EvaluationCLI, the
  CopilotConnectorSkill repo, and the TypeScript enhancer.
* Probe order locked to 5 entries matching the TS source (test asserts
  the literal order).
* Semver compare ported with the TS quirks (3-component only, 'v'
  prefix, non-numeric segment -> 0).
* ToolResolverOptions exposes WorkflowRepoRoot + SrcRoot with
  CCW_WORKFLOW_REPO / CCW_SRC_ROOT env fallbacks since C# has no
  __dirname.
* CheckNodeMinimumAsync spawns 'node -v' (no in-process equivalent
  of TS process.versions.node).
* GetUserHomeDirectory mirrors Node os.homedir(): prefer USERPROFILE
  env var so test isolation works on Windows where SHGetKnownFolderPath
  ignores env overrides.
* 11 tests including fallback-path isolation via per-test USERPROFILE
  override + temp dirs.

Reviewer fixes  GPT-5.5 BLOCKERs
* MetadataProvenance.{titleFromSource,urlFromSource,iconUrlFromSource}
  changed int -> double. TS computes round3(count/total) = 0..1 fraction;
  int would silently truncate every provenance value to 0 on round-trip
  from a real Node ScoredReport. Comment in Job.cs cites the review.
* ProcessRunner.cs rewritten for the concurrent-pump race:
  - SemaphoreSlim write gate serializes all StringBuilder appends,
    log-file writes, and channel writes across the stdout/stderr pumps.
  - Channel SingleWriter switched to false (two pumps = two writers).
  - sink.TryComplete() moved into outer finally so UI consumers
    unblock on every exit path, including cancellation.
  - Null Env values now Remove(key) from the child env (matches Node
    spawn({env:{KEY: undefined}}) semantics).
  - Cancellation drains pump tasks via SafeAwait after Kill, with a
    bounded 5-second OS-reap grace via WaitForExitAsync().WaitAsync.
  - CA2016 fixed by passing CancellationToken.None explicitly.
* Templater .hbs extension checks switched from OrdinalIgnoreCase to
  Ordinal  Node endsWith('.hbs') is case-sensitive so '.HBS' files
  must copy verbatim, not render.

Reviewer fixes  Opus-4.8 BLOCKERs/IMPORTANTs
* CanonicalHash: BLOCKER B2  List<T>.Sort is unstable; V8
  Array.prototype.sort is stable (ES2019+). Swapped both
  entries.Sort(...) callsites in HashDataset + HashEvalSet to
  stable LINQ OrderBy. Without this, two file names that collate
  equal under InvariantCulture (e.g. 'ab' vs 'a\u00ADb') could
  manifest-order differently across runs and break the SHA-256.
* Templater: IMPORTANT I3  File.ReadAllText auto-strips a UTF-8 BOM;
  Node fs.readFileSync(p,'utf-8') preserves it. Switched .hbs read to
  raw bytes + UTF8Encoding(false) GetString so BOM'd templates render
  byte-identically.
* ProcessRunner: IMPORTANT I2  CreateLogChannel default mode changed
  Wait -> DropOldest. A stalled UI consumer must never block the child
  process; the lossless source of truth is the log file (always
  written) and RunResult.Output. Optional fullMode parameter still
  lets tests opt into Wait when they need backpressure semantics.

New tests
* CanonicalHashParityTests.HashDataset_StringsTiedUnderInvariantCulture_PreservesInputOrder
   tripwire for any future regression to unstable sort.
* TemplaterTests.RenderFileToDir_HbsWithUtf8Bom_PreservesBomInOutput
  + RenderFileToDir_VerbatimWithUtf8Bom_PreservesBomInOutput  pin
  the BOM contract for both render and copy paths.
* TemplaterTests.RenderFileToDir_HbsExtensionCaseSensitive_HBSTreatedAsVerbatim
  + sibling lowercase assertion  pin .hbs case-sensitivity.
* ProcessRunnerTests.RunAsync_EnvNullValue_RemovesVariableFromChild
   uses cmd 'IF DEFINED' for a reliable present/absent distinction.
* ProcessRunnerTests.RunAsync_Cancellation_CompletesLogSink  proves
  TryComplete in the cancellation path.
* JobJsonRoundtripTests.MetadataProvenance_FractionalFields_RoundtripWithoutTruncation
   locks the int->double change.

All 67 Ccw.Core.Tests pass.

+++ winui3/src/Ccw.Core/Models/Job.cs
@@ -84,9 +84,14 @@ public sealed record ScoreConfig
 /// <summary>Metadata provenance counts. TS: anonymous nested type in <c>ScoredReport.metadataProvenance</c>.</summary>
 public sealed record MetadataProvenance
 {
-    [JsonPropertyOrder(0)] public required int TitleFromSource { get; init; }
-    [JsonPropertyOrder(1)] public required int UrlFromSource { get; init; }
-    [JsonPropertyOrder(2)] public required int IconUrlFromSource { get; init; }
+    // BLOCKER fix from GPT review: TS `number` here is a 0..1 fraction
+    // (round3(itemsWithSourceTitle / itemCount) in identity-transform.ts),
+    // not an integer. Using `int` would silently truncate every
+    // metadata-provenance count to 0 on round-trip from a real Node
+    // ScoredReport.
+    [JsonPropertyOrder(0)] public required double TitleFromSource { get; init; }
+    [JsonPropertyOrder(1)] public required double UrlFromSource { get; init; }
+    [JsonPropertyOrder(2)] public required double IconUrlFromSource { get; init; }
     [JsonPropertyOrder(3)] public required int SchemaPropertiesPromotedToSearchable { get; init; }
     [JsonPropertyOrder(4)] public required int SchemaPropertiesPromotedToRefinable { get; init; }
 }
